// src/apps/frontend/src/app/api/model/user-login-response-dto.ts
/**
 * ToDo Application API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserLoginResponseDto { 
    /**
     * ユーザーのUUID
     */
    id: string;
    /**
     * ユーザーのメールアドレス
     */
    email: string;
    /**
     * アクセストークン
     */
    accessToken: string;
}