// src/apps/frontend/src/app/api/model/user-get-response-dto.ts
/**
 * ToDo Application API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserGetResponseDto { 
    /**
     * ユーザーのUUID
     */
    id: string;
    /**
     * ユーザー名
     */
    name?: string;
    /**
     * ユーザーのメールアドレス
     */
    email: string;
    /**
     * 作成日時
     */
    createdAt?: string | null;
    /**
     * 更新日時
     */
    updatedAt?: string | null;
}